[CLI] removes 'CLI'
This was causing some minor headaches. Namely, that having the the
executable target defined in `Package.swift` would generate a Xcode
project file which wouldn't compile for _iOS_.

The perfect solution would be a folder named **Examples** sitting
at the root, and this would have sub-folders for each example
executable I want. I tried this, but SPM couldn't resolve the
`WalletKit` framework correctly (as it appeard as `../..' in the
dependencies list).